Endelig versjon som leveres 2. oktober 2017

+++ src/main/java/Innlevering1/Database/TableManager.java
@@ -22,6 +22,10 @@ public class TableManager {
             statement.executeUpdate("DROP TABLE " + tableName);
             return "Table " + tableName + " is dropped!";
         }catch (SQLException e){
+            if (e.getErrorCode() == 1217)
+                return "Can't delete table because it is " +
+                        "connected to another table. Restart the Application" +
+                        " if you want to delete this table.";
             return "Could not drop table";
         }
     }
